:recycle: Removed overkill class
Really don't need a separate class that just contains a list of ints -
easier to just pass them into a list of ints and then have some
extension methods on that.

Still not ideally performant - would be better to stop at the point
where an error is found, and then try removing that level, the one
below, and the one above. But that becomes a lot less readable.