:sparkles: Solve day 02, 2024 (#8)
* :construction: Add template code for day 2

Empty tests and an empty class.

* :test_tube: Add test for part 1

Currently failing because nothing is implemented.

* :sparkles: Solve part 1

* :test_tube: Add test for part 2

* :poop: Messy version that works

Terrible method calling and hierarchy but gets the job done for now.

* :recycle: Removed overkill class

Really don't need a separate class that just contains a list of ints -
easier to just pass them into a list of ints and then have some
extension methods on that.

Still not ideally performant - would be better to stop at the point
where an error is found, and then try removing that level, the one
below, and the one above. But that becomes a lot less readable.