:sparkles: Solve day 07 2024 (#14)
* :construction: Generate template code

* :test_tube: Add test for part 1

* :construction: Parse input for part 1

Made the parser method public - this involved changing in all the other
classes but means I was able unit test the parsing.

* :test_tube: Add test for combination method

* :construction: Implement combination method

Method using binary to get all combinations - at the moment will only
work for valid bases in the `Convert` to string method - can easily be
extended.

* :sparkles: Solve part 1

Had to switch to using longs because the numbers got too big.

* :recycle: Tidy into methods a bit

Not perfect with the for loop logic but at least has a few methods to
break things out.

* :test_tube: Add test for part 2

* :sparkles: Solve part 2

Not super fast - could be improved (I think) by caching the combinations
but it does work...

* :recycle: Move combinations to common

This can be used again in future and can be unit tested outside of the
day 7 context.