:sparkles: Solve day 08 2024 (#15)
* :construction: Add template code

* :test_tube: Add test for part 1

* :construction: Parse input for Part 1

Also tests this input parsing

Removed an unused import while I was at it from the `Coordinate`
class I'm using.

* :test_tube: Test `GetAntinodes` method

Pretty central to the solution logic...

* :construction: Implement `GetAntinodes` method

* :construction: Adds combinations method

Already have a static class with useful methods for getting
combinations, so added to that one.

Adds a unit test of it.

* :sparkles: Solve part 1

* :test_tube: Add test for part 2

* :recycle: Use vectors

Use a vector to get the next antinodes.

Will make it easier for part 2.

* :sparkles: Solve part 2